+++ frontend/beCompliant/src/api/apiConfig.ts
@@ -1,14 +1,12 @@
 // Query Keys / Individual paths
 const PATH_USERINFO = '/userinfo';
 const PATH_CONTEXTS = '/contexts';
-const PATH_DUMP_CSV = '/dump-csv';
 
 // Base URLs
 
 const API_URL_BASE = import.meta.env.VITE_BACKEND_URL;
 // API Endpoints
 const API_URL_CONTEXTS = `${API_URL_BASE}${PATH_CONTEXTS}`;
-const API_URL_DUMP_CSV = `${API_URL_BASE}${PATH_DUMP_CSV}`;
 export const API_URL_USERINFO = `${API_URL_BASE}${PATH_USERINFO}`;
 
 export const apiConfig = {
@@ -19,10 +17,12 @@ export const apiConfig = {
       queryKey: (contextId: string) => [PATH_CONTEXTS, contextId],
       url: (contextId: string) => `${API_URL_CONTEXTS}/${contextId}`,
     },
+
     forTeam: {
       queryKey: (teamId: string) => [PATH_CONTEXTS, teamId],
       url: (teamId: string) => `${API_URL_CONTEXTS}?teamId=${teamId}`,
     },
+
     forIdAndTeam: {
       queryKey: (contextId: string, teamId: string) => [
         PATH_CONTEXTS,
@@ -31,6 +31,7 @@ export const apiConfig = {
       ],
       url: (contextId: string) => `${API_URL_CONTEXTS}/${contextId}`,
     },
+
     forTeamAndForm: {
       queryKey: (teamId: string, formId: string) => [
         PATH_CONTEXTS,
@@ -40,9 +41,5 @@ export const apiConfig = {
       url: (teamId: string, formId: string) =>
         `${API_URL_CONTEXTS}?teamId=${teamId}&formId=${formId}`,
     },
-  },
-  dumpCSV: {
-    queryKey: [PATH_DUMP_CSV],
-    url: API_URL_DUMP_CSV,
   },
 } as const;
