+++ frontend/beCompliant/src/pages/frontPage/FrontPage.tsx
@@ -6,14 +6,12 @@ import {
   Button,
   Skeleton,
   StackSeparator,
-  toaster,
 } from '@kvib/react';
 import { Page } from '../../components/layout/Page';
 import { useUser } from '../../hooks/useUser';
-import { apiConfig } from '../../api/apiConfig';
-import { axiosFetch } from '../../api/Fetch';
 import RedirectBackButton from '../../components/buttons/RedirectBackButton';
 import TeamContexts from './TeamContexts';
+import { handleExportCSV } from '../../utils/csvExportUtils';
 
 export default function FrontPage() {
   const {
@@ -48,48 +46,6 @@ export default function FrontPage() {
         </Center>
       </>
     );
-  }
-
-  async function handleExportCSV() {
-    try {
-      const response = await axiosFetch<Blob>({
-        url: apiConfig.dumpCSV.url,
-        method: 'GET',
-        responseType: 'blob',
-      });
-
-      if (!response.data) {
-        throw new Error('No data received for CSV');
-      }
-
-      const blob = new Blob([response.data], {
-        type: 'text/csv;charset=utf-8;',
-      });
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement('a');
-      link.href = url;
-      link.setAttribute('download', 'data.csv');
-
-      document.body.appendChild(link);
-      link.click();
-
-      document.body.removeChild(link);
-
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      console.error('Error downloading CSV:', error);
-      const toastId = 'export-csv-error';
-      if (!toaster.isVisible(toastId)) {
-        toaster.create({
-          id: toastId,
-          title: 'Å nei!',
-          description:
-            'Det kan være du ikke har tilgang til denne funksjonaliteten:',
-          type: 'error',
-          duration: 5000,
-        });
-      }
-    }
   }
 
   return (
